python/requirements.txt
requests==2.31.0
firebase-admin==6.2.0
python-dotenv==1.0.0
newsdataapi